kpatch: rmmod module of the same name before loading a module
With klp->replace, when a patch is replaced, it is no longer an active
patch, but the module is still loaded with a refcount of 0. If the user
tries to load the patch again, insmod will fail with EEXIT. To avoid
such errors, run a proactive rmmod before loading the module. This is a
no-op if the module is not loaded or is actually in use.

Also, update module_ref_count() to only succeed with refcnt > 1.

Signed-off-by: Song Liu <song@kernel.org>