add cross-compilation support
This patch introduces new option for kpatch-build
script "--cross-compile" which can be used for
specifying cross-complier prefix.
It allows to build live patches not only on
target system, but also on hosts for a target other
than the one on which the compiler is running.

This patch also adds a new --distro option which
introduces ability to override distro name, for
example, during cross-compilation build when
livepatch modules will be ran on the target
which differs from host.

Also removed quotes in exec lines, so it is
possible to pass multi-component strings like
"ccache x86_64-xelinux-linux-" as cross-compiler

Signed-off-by: Ruslan Bilovol <rbilovol@cisco.com>
Signed-off-by: Oleh Matiusha <oleh.matiusha@globallogic.com>