test: Add data relocation test
This patch adds a test case that the klp-module has a
`.klp.rela.<module name>..rodata.<...>` section. This pattern can be an
error if architecture-specific relocation in the kernel does not fully
support Livepatch.

The prerequisite for this test is that CONFIG_USB_SERIAL_IR=m and
CONFIG_USB_SERIAL_WWAN=m are set. In the patched-modules,
ir-usb.ko's static local variables refers to usb_wwan.ko's symbols,
so `.klp.rela.ir-usb..rodata.<...>` section are generated into
test-data-reloc.ko.

Signed-off-by: Keiya Nobuta <nobuta.keiya@fujitsu.com>