kpatch-build: add RHEL -debug kernel support
For RHEL distribution, if kpatch-build fetches the current kernel source
rpm, determine kernel/kernel-debug variant based on the ARCHVERSION,
(does it end in $(arch)+debug).  If the source rpm is provided to
kpatch-build, look for new command line option, --rhel-debug-config,
which will indicate to use the debug kernel configuration from the
source rpm.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>