kpatch-build: let locals_match to match 90% of local symbols
With clang PGO is used, the compiler uses function signature to match
functions to profile data. When the signature of a function changed as
part of livepatch, the profile data is ignore. This may cause the compiler
to make different inline decisions, and thus cause mismatch in local
functions.

To adapt to this case, let locals_match to count the number of match and
mismatch symbols. If 90% of the symbols match, we consider the two files
are the same.

Signed-off-by: Song Liu <song@kernel.org>