kpatch-build: relax vmlinux "Linux version" grep
When determining the linux version from a vmlinux file, we assume that
`strings` output will show "Linux version <version>" at the beginning of
the line.

This assumption was observed to not hold true for clang on s390x:

  $ strings ./build/linux-5.18/vmlinux | grep -m1 -e "Linux version"
  mY*Linux version 5.18.0 (root@xyz.com) (clang version 14.0.5 (Fedora 14.0.5-1.fc36), GNU ld version 2.37-36.fc36) #7 SMP Wed Oct 19 12:47:00 EDT 2022

Relax the regular expression to look for "Linux version <word>", but not
anchored at the beginning of the line.

Reported-by: Yulia Kopkova <ykopkova@redhat.com>
Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>