fix up wrong rela of debug sections
if one debug section's rela reference to __verbose section's symbol
(for example, pr_debug) and some of __verbose section's symbols are
included,then the __verbose section will be included.
But other symbols of __verbose may be not included,then the debug
section's relas reference to the not included symbols would result
into bad rela in the end output elf file.
So if the rela's symbol is not included,just delete that rela.

Signed-off-by: Hongchen Zhang <zhanghongchen@loongson.cn>