use dynamic prefix offset instead of hardcoding
In commit [1], kpatch added support for function padding,
and CONFIG_CFI_CLANG, which hardcoded a value of 16 for
the prefix size.

In some cases, the padding around __cfi prefixed functions can
vary. For example, in Photon OS 5.0, the __cfi prefix
size is modified in a patch for the gcc RAP plugin [2].

Since we have read the prefix size anyways, we can use it
instead of hardcoding.

Ref:
1. https://github.com/dynup/kpatch/commit/3e54c63b175b68cf48654c119e62bda398d0c018
2. https://github.com/vmware/photon/blob/5.0/SPECS/linux/secure/gcc-rap-plugin-with-kcfi.patch

Signed-off-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>