CDO: Introduce log_error()
log_normal() is used in some critical errors, such as:

"changed section XXXX not selected for inclusion".

It is not very obvious this is a fatal error.

Introduce log_error() and use it for critical messages.

Note that we are not using ERROR() here, because ERROR() will terminate
CDO immediately, which is not the desired behavior.

Signed-off-by: Song Liu <song@kernel.org>
Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com> [merge conflict]