create-diff-object: allow ignore not exist section
When we adjust one patch with KPATCH_IGNORE_SECTION,
different arch may different building action which
may lead to this section not created under some arch.

For example:
```
create-diff-object: ERROR: inode.o: kpatch_mark_ignored_sections: 2993: KPATCH_IGNORE_SECTION: can't find .rela.return_sites
```

The ignored section is what we dont need, so, if this
section is not exist, it should not do much effect
to us since we try to throw it away. We should allow
it continue to build.

Signed-off-by: Wardenjohn <zhangwarden@gmail.com>