create-diff-object: extract and retain symbol namespace from .modinfo
Kernel module namespace imports are saved in the .modinfo section in
key-value strings.  In order for a livepatch to access those symbols in
those namespaces, it needs to provide the same imports else modpost
refuses to process it.  For example:

  ERROR: modpost: module livepatch-test uses symbol dma_buf_export from namespace DMA_BUF, but does not import it.

We don't need the entire .modinfo section, as it carries other key-value
pairs that we are not interested in.  Selectively extract all of the
"import_ns=<value" strings and then restore them in the ELF output for
inclusion in the livepatch .ko.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>