test/unit/objs: Update reference for arm64 objs
AArch64 test objects are now available in the kpatch-unit-test-objs
repository with commit 4ad64a06b6d0 ("Merge pull request #54 from
dynup/aarch64")

Update the submodule reference to the above commit so the objects are
available to the test suite. Add aarch64 to the unit Makefile so the
tests are run.

Signed-off-by: Puranjay Mohan <pjy@amazon.com>